+++ app/src/api/api.js
@@ -349,7 +349,8 @@ export const mapContributionDataToForm = contribution => {
     subTypeOfContribution: subtype,
     typeOfContributor: contributorType,
     inKindType,
-    oaeType: DataToOaeTypeTypeFieldMap.get(oaeType),
+    oaeType,
+    // DataToOaeTypeTypeFieldMap.get(oaeType),
     amountOfContribution: amount,
     checkNumber,
     submitForMatch: submitForMatch ? 'Yes' : 'No',
